fix(api): moves unit usecase tests to integration
